dev(feat): new object TokenService

+++ src/main/kotlin/dto/LoginResponse.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 data class LoginResponse(
     val accessToken: String,
     val tokenType: String,
-    val expiresIn: Int // seconds
+    val expiresIn: Long // seconds
 )
